+++ engine/src/GuiTextfield.cpp
@@ -41,7 +41,8 @@ void GuiTextfield::Draw(sf::RenderTarget* target) {
 	else
 		mText.SetString(mCaption);
 
-	mText.SetPosition(mPosition.x+8, mPosition.y + mDimension.y / 2 - mText.GetRect().Height / 2);
+	mText.SetPosition(mPosition.x+8,
+					  int(mPosition.y + mDimension.y / 2 - mText.GetRect().Height / 2)); // convert to int to prevent blurry font due to half pixel coordinates
     target->Draw(mText);
 
     // draw cursor
